adding more documentation and github workflows, moving tests around
