+++ .github/workflows/rust_pr_checks.yml
@@ -1,0 +1,45 @@
+name: Rust PR Checks
+
+on:
+  pull_request:
+    branches: [ "master" ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Install stable toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: stable
+          override: true
+      - name: Build
+        run: cargo build --verbose
+      - name: Check
+        run: cargo check
+      - name: Run tests
+        run: cargo test --verbose --all --all-features
+      - name: Lint publish
+        run: cargo publish --dry-run --color always --all-features --frozen --locked
+
+  bump:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Install cargo-bump
+        run: cargo install cargo-bump
+
+      - name: Bump Cargo Runner Version
+        run: cargo bump minor
+        working-directory: ./cargo-openapi-generator
+      - name: Bump Cargo Core Version
+        run: cargo bump minor
+        working-directory: ./core
+      - name: Bump Generator Version
+        run: cargo bump minor
+        working-directory: ./generator
+
+      - name: Push new version
+        run: git push
