+++ .github/workflows/rust_publish.yml
@@ -1,0 +1,39 @@
+name: Rust Publish
+
+on:
+  push:
+    branches: [ "master" ]
+
+env:
+  CRATES_TOKEN: ${{ secrets.CRATES_SECRET }}
+
+jobs:
+  publish:
+    name: Publish
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+      - name: Install stable toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: stable
+          override: true
+
+      - name: Fetch version
+        id: ver
+        uses: dante-signal32/rust-app-version@v1.2.0
+        with:
+          cargo_toml_folder: core/
+      - name: Create Release Tag
+        uses: rickstaa/action-create-tag@v1
+        with:
+          tag: ${{ steps.ver.outputs.app_version }}
+
+      - name: Publish Cargo Runner
+        run: cargo publish --package cargo-openapi --token ${CRATES_TOKEN}
+      - name: Publish Cargo Core
+        run: cargo publish --package sw4rm-rs --token ${CRATES_TOKEN}
+      - name: Publish Cargo Runner
+        run: cargo publish --package sw4rm-rs-generation --token ${CRATES_TOKEN}
