+++ core/src/reference.rs
@@ -126,29 +126,4 @@ impl TryFrom<String> for Reference {
 #[cfg(test)]
 mod tests {
     use super::Reference;
-
-    #[test]
-    fn test_regex_captures_swagger_yaml() {
-        let hay = "#/parameters/nested_response";
-        let path = Reference::try_from(hay.to_string()).unwrap();
-        println!("{path:?}")
-    }
-    #[test]
-    fn test_regex_captures_swagger_json() {
-        let hay = "#/definitions/account-v1.AccountDto";
-        let path = Reference::try_from(hay.to_string()).unwrap();
-        println!("{path:?}")
-    }
-    #[test]
-    fn test_regex_captures_openapi_yaml() {
-        let hay = "#/components/schemas/Pets";
-        let path = Reference::try_from(hay.to_string()).unwrap();
-        println!("{path:?}")
-    }
-    #[test]
-    fn test_regex_captures_openapi_json() {
-        let hay = "#/components/schemas/account-v1.AccountDto";
-        let path = Reference::try_from(hay.to_string()).unwrap();
-        println!("{path:?}")
-    }
 }
